fix message & user

+++ internal/domain/message.go
@@ -1,13 +1,13 @@
 package domain
 
 type Message struct {
-	Id       int64
+	Id       int
 	SenderId int64
 	Date     string
 	Text     string
 }
 
-func NewMessage(id int64, senderId int64, date string, text string) *Message {
+func NewMessage(id int, senderId int64, date string, text string) *Message {
 	return &Message{
 		Id:       id,
 		SenderId: senderId,
